# backend/docs/swagger.json
{
    "components": {
        "schemas": {
            "database.Project": {
                "properties": {
                    "color": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "order": {
                        "type": "integer"
                    },
                    "tasks": {
                        "items": {
                            "$ref": "#/components/schemas/database.Task"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "database.Task": {
                "properties": {
                    "completed_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "due": {
                        "type": "string"
                    },
                    "end_datetime": {
                        "type": "string"
                    },
                    "google_event_id": {
                        "type": "string"
                    },
                    "has_time": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "labels": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "order": {
                        "type": "integer"
                    },
                    "project": {
                        "$ref": "#/components/schemas/database.Project"
                    },
                    "project_id": {
                        "type": "integer"
                    },
                    "recurrence": {
                        "type": "string"
                    },
                    "reminders": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "start_datetime": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "main.CreateProjectRequest": {
                "properties": {
                    "color": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "order": {
                        "type": "integer"
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object"
            },
...
            "main.FindResponse": {
                "properties": {
                    "count": {
                        "type": "integer"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/main.SearchResult"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "main.SearchResult": {
                "properties": {
                    "content": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "subtitle": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "main.UpdateProjectRequest": {
                "properties": {
                    "color": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "order": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "Task management REST API",
        "title": "Dimaist API",
        "version": "1.0"
    },
...
    "paths": {
        "/find": {
            "get": {
                "operationId": "find_items",
                "parameters": [
                    {
                        "description": "Search query",
                        "in": "query",
                        "name": "q",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.FindResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Search across tasks and projects",
                "tags": [
                    "search"
                ]
            }
        },
        "/projects": {
            "get": {
                "operationId": "list_projects",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/database.Project"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List all projects",
                "tags": [
                    "projects"
                ]
            },
            "post": {
                "operationId": "create_project",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/main.CreateProjectRequest",
                                        "summary": "project",
                                        "description": "Project to create"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Project to create",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/database.Project"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Create a new project",
                "tags": [
                    "projects"
                ]
            }
        },
        "/projects-reorder": {
            "put": {
                "operationId": "reorder_projects",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "items": {
                                            "type": "integer"
                                        },
                                        "title": "project_ids",
                                        "type": "array"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Ordered list of project IDs",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Reorder projects",
                "tags": [
                    "projects"
                ]
            }
        },
        "/projects/{project_id}": {
            "delete": {
                "operationId": "delete_project",
                "parameters": [
                    {
                        "description": "Project ID",
                        "in": "path",
                        "name": "project_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Delete a project",
                "tags": [
                    "projects"
                ]
            },
            "put": {
                "operationId": "update_project",
                "parameters": [
                    {
                        "description": "Project ID",
                        "in": "path",
                        "name": "project_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/main.UpdateProjectRequest",
                                        "summary": "project",
                                        "description": "Project fields to update"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Project fields to update",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Update a project",
                "tags": [
                    "projects"
                ]
            }
        },
        "/projects/{project_id}/tasks/reorder": {
            "put": {
                "operationId": "reorder_project_tasks",
                "parameters": [
                    {
                        "description": "Project ID",
                        "in": "path",
                        "name": "project_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "items": {
                                            "type": "integer"
                                        },
                                        "title": "task_ids",
                                        "type": "array"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Ordered list of task IDs",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Reorder tasks within a project",
                "tags": [
                    "projects"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:3000/"
        }
    ]
}